libs/architect-observability/pyproject.toml
[project]
name = "architect-observability"
version = "0.1.0"
description = "OpenTelemetry tracing and Prometheus metrics for ARCHITECT services"
requires-python = ">=3.12"
dependencies = [
    "opentelemetry-api>=1.27",
    "opentelemetry-sdk>=1.27",
    "opentelemetry-exporter-otlp-proto-grpc>=1.27",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.48b0",
    "opentelemetry-instrumentation-redis>=0.48b0",
    "opentelemetry-instrumentation-httpx>=0.48b0",
    "prometheus-fastapi-instrumentator>=7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/architect_observability"]
